requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
feedparser>=6.0.11
python-dotenv>=1.0.1
Pillow>=10.0.0
requests>=2.32.0
pytz>=2024.1
